Backend general terminado

+++ backend/src/main/java/com/grindsup/backend/repository/AgendaRepository.java
@@ -3,5 +3,5 @@ package com.grindsup.backend.repository;
 import com.grindsup.backend.model.Agenda;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgendaRepository extends JpaRepository<Agenda, Integer> {
+public interface AgendaRepository extends JpaRepository<Agenda, Long> {
 }
