+++ backend/src/main/java/com/grindsup/backend/repository/AlumnoRepository.java
@@ -3,5 +3,5 @@ package com.grindsup.backend.repository;
 import com.grindsup.backend.model.Alumno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlumnoRepository extends JpaRepository<Alumno, Integer> {
+public interface AlumnoRepository extends JpaRepository<Alumno, Long> {
 }
