+++ backend/src/main/java/com/grindsup/backend/repository/EjercicioRepository.java
@@ -3,5 +3,5 @@ package com.grindsup.backend.repository;
 import com.grindsup.backend.model.Ejercicio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EjercicioRepository extends JpaRepository<Ejercicio, Integer> {
+public interface EjercicioRepository extends JpaRepository<Ejercicio, Long> {
 }
