+++ backend/src/main/java/com/grindsup/backend/repository/EstadoRepository.java
@@ -3,5 +3,5 @@ package com.grindsup.backend.repository;
 import com.grindsup.backend.model.Estado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadoRepository extends JpaRepository<Estado, Integer> {
+public interface EstadoRepository extends JpaRepository<Estado, Long> {
 }
