+++ backend/src/main/java/com/grindsup/backend/repository/PlanEntrenamientoRepository.java
@@ -3,5 +3,5 @@ package com.grindsup.backend.repository;
 import com.grindsup.backend.model.PlanEntrenamiento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanEntrenamientoRepository extends JpaRepository<PlanEntrenamiento, Integer> {
+public interface PlanEntrenamientoRepository extends JpaRepository<PlanEntrenamiento, Long> {
 }
