+++ backend/src/main/java/com/grindsup/backend/repository/RutinaRepository.java
@@ -3,5 +3,5 @@ package com.grindsup.backend.repository;
 import com.grindsup.backend.model.Rutina;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RutinaRepository extends JpaRepository<Rutina, Integer> {
+public interface RutinaRepository extends JpaRepository<Rutina, Long> {
 }
