+++ backend/src/main/java/com/grindsup/backend/repository/TurnoRepository.java
@@ -3,5 +3,5 @@ package com.grindsup.backend.repository;
 import com.grindsup.backend.model.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurnoRepository extends JpaRepository<Turno, Integer> {
+public interface TurnoRepository extends JpaRepository<Turno, Long> {
 }
